+++ app/Http/Controllers/GuruController.php
@@ -56,10 +56,16 @@ class GuruController extends Controller
                 if ($request->id !== null) {
                     # code...
                     $datas = Guru::find($request->id);
-                    $image_path1 = public_path("guru_image\\".$datas->guru_image);
-                    $image_path2 = public_path("image_guru\\".$datas->guru_image);
-                    unlink($image_path1);
-                    unlink($image_path2);
+                    $image_path1 = public_path('guru_image/'.$datas->guru_image);
+                    $image_path2 = public_path('image_guru/'.$datas->guru_image);
+                    if (File::exist($image_path1)) {
+                        # code...
+                        unlink($image_path1);
+                    }
+                    if (File::exist($image_path2)) {
+                        # code...
+                        unlink($image_path2);
+                    }
                 }
 
                 $filename    = time().'.'.$request->guru_image->getClientOriginalExtension();
@@ -103,10 +109,16 @@ class GuruController extends Controller
     public function remove_guru(Request $request)
     {
         $data = Guru::find($request->id);
-        $image_path1 = public_path("guru_image\\".$data->guru_image);
-        $image_path2 = public_path("image_guru\\".$data->guru_image);
-        unlink($image_path1);
-        unlink($image_path2);
+        $image_path1 = public_path("guru_image/".$data->guru_image);
+        $image_path2 = public_path("image_guru/".$data->guru_image);
+        if (File::exists($image_path1)) {
+            # code...
+            unlink($image_path1);
+        }
+        if (File::exists($image_path2)) {
+            # code...
+            unlink($image_path2);
+        }
         $data->delete();
         # code...
         return response()->json(
